# Performance improvement for NcWriter DateTime conversions