--- a/src/lib-python/ioda_conv_ncio.py
+++ b/src/lib-python/ioda_conv_ncio.py
@@ -1,12 +1,13 @@
 #!/usr/bin/env python
 
-from __future__ import print_function
+import datetime as dt
+import math
+import sys
+from collections import OrderedDict
+
 import numpy as np
 import netCDF4
 from netCDF4 import Dataset
-import datetime as dt
-import sys
-from collections import OrderedDict
 
 ###################################################################################
 # SUBROUTINES
@@ -183,18 +184,17 @@ class NcWriter(object):
         Output: numpy dtype:'f4' array of time offsets in decimal hours from the reference date.
         """
         offset = np.zeros((self._nlocs), dtype='f4')
-        ref_offset = self._ref_date_time.hour + self._ref_date_time.minute/60. + self._ref_date_time.second/3600. #decimal hours
-        date_offset_cache = {} #Cache a map from bytes[0:10] to the corresponding date's offset from reference date
+        ref_offset = math.floor(self._ref_date_time.hour*3600. + self._ref_date_time.minute*60. + self._ref_date_time.second)/3600.  # decimal hours
+        date_offset_cache = {}  # Cache a map from bytes[0:10] to the corresponding date's offset from reference date
         for i in range(len(datestr_vec)):
             date_str = datestr_vec[i][0:10].tobytes()
             date_offset = date_offset_cache.get(date_str)
             if date_offset is None:
-                #compute and cache the date's offset.
-                #There are at most 2 distinct dates per cycle for any window length < 24hrs
-                date_offset = (self._ref_date_time.date() - dt.date.fromisoformat(date_str.decode('ascii'))).total_seconds()/3600.
+                # There are at most 2 distinct dates per cycle for any window length < 24hrs
+                date_offset = (dt.date.fromisoformat(date_str.decode('ascii')) - self._ref_date_time.date()).total_seconds()/3600.
                 date_offset_cache[date_str] = date_offset
             t = dt.time.fromisoformat(datestr_vec[i][11:19].tobytes().decode('ascii'))
-            t_offset = t.hour + t.minute/60. + t.second/3600.
+            t_offset = math.floor(t.hour*3600. + t.minute*60. + t.second)/3600.
             offset[i] = date_offset + (t_offset - ref_offset)
         return offset