@CoryMartin-NOAA 2020-08-11
@srherbener am I wrong, or was there some discussion before about not even using the time@MetaData anymore in IODA and only using the string timestamps? If that's the case, can we remove this from the converter API?