@markjolah 2020-08-11
Note.  Apparently there is no `.fromisoformat()` in Python 3.6.  I will have to figure out a workaround for the older python support.

Not sure about what the overall plans for the variables... I actually prefer the format of `time@MetaData`, as it is more efficient in both space and time.  Working with strings is always slower, and we we're running into that as a bottleneck with plotting HofX output.