@CoryMartin-NOAA 2020-08-11
The problem is we don't want the time to be stored as reference to some value. GSI does this and it's fine for now, in reference to analysis time, but long term the actual time should be stored. However, the Unix epoch time might be a good solution to use in time@MetaData? But this is probably going to have to depend on IODA-engines.