@markjolah 2020-08-11
Yeah, I definitely see there are +/- for each option.  From a scientific and user-practicality standpoint I think the textual representation is the best, and that's why it will end up being the final choice.   However, for performance, the floating point representation will always be way better and by about a factor of 10-1000 depending on the situation.   We'll just have to work to get that closer to the lower end of that range with tricks like employed here.