@markjolah 2020-08-11
Ok. Fixed with an even more efficient version that has no python version problems.  Now the time in `NcWriter::ConvertToTimeOffset` is down to: `0.211 s / 0.973 s = 21.7%` of total and the overall speedup is a factor of 2.9 for the 45mb dataset.  Similar code might be useful in other areas as profiling determines.