@srherbener 2020-08-12
@CoryMartin-NOAA, you are correct that the Reference/Offset style of datetime is going away. This code is a leftover from when we only could handle the Reference/Offset style in the IODA reader, and were transitioning to handling the ISO 8601 formatted strings in the IODA reader. That has been done, so this step in the converters can be deleted.

The plan is to move forward with absolute datetime information in ISO 8601 formatted strings. If we run into performance issues with strings, we may want to consider using something like two integers (YYYYMMDD, hhmmss for example) instead. The target is to get the datetime information into an oops DateTime object so we can add a new constructor if we decide to change the way we store datetime info. The nice thing about ISO 8601 strings is that they are an accepted standard, readable and easily understood.

This speedup modification is great work, but I would like to caution that we are in a time where the ioda converters are going to change (transition to the new interface) and the probability of optimizing code that is going to be thrown away is higher than usual. I recommend that we wait until after the migration to the new interface to continue with performance optimization.